FLUME-2698 - Upgrade Jetty Version
updated patch based on the patch provided by Tristan Stevens

https://reviews.apache.org/r/52955

This patch upgrades the HTTPSource and HTTP Metrics server to Jetty 9.4. It also adds the ability to further configure the Jetty server and also run in multi-threaded mode. Finally, registered Jetty with the MBeans server to allow fine-grained performance tuning/monitoring if required.